Added ContextMenu to MVDs
I am in the middle of factoring out the Configuration Pages, replacing
these with a ContextMenu that drops down when a View Delegate is
clicked.
However the ContextMenus are not yet displayed properly
Note that currently the actions on the ContextMenus don't actually do
anything, that will come as soon as get the ContextMenus displaying
properly!